+++ lib/microverse_twitterati/retweet.rb
@@ -1,5 +1,5 @@
 require 'dotenv/load'
-require 'twitter'
+require_relative './initiator'
 
 module MicroverseTwitterati
   class Retweet
@@ -8,7 +8,7 @@ module MicroverseTwitterati
     attr_reader :client
 
     def initialize
-      @client = setup_client
+      @client = Initiator.new.client
       retweet!
     end
 
@@ -24,15 +24,6 @@ module MicroverseTwitterati
 
     def bot_followers
       client.follower_ids
-    end
-
-    def setup_client
-      Twitter::REST::Client.new do |config|
-        config.consumer_key = ENV['TWITTER_API_KEY']
-        config.consumer_secret = ENV['TWITTER_SECRET_KEY']
-        config.access_token = ENV['TWITTER_ACCESS_TOKEN']
-        config.access_token_secret = ENV['TWITTER_TOKEN_SECRET']
-      end
     end
   end
 end
